feat: add initial rusEFI support
- Updated integration tests to handle ECU disconnection gracefully, ensuring commands fail as protocol errors instead of crashing.
- Introduced a new test suite for rusEFI, validating communication and functionality with a simulated rusEFI ECU.
- Added socket link tests to verify behavior on TCP reset scenarios.
- Improved write path tests to ensure correct handling of burn commands based on ECU definitions.
- Enhanced fake ECU implementation to support both Speeduino and rusEFI protocols, allowing for flexible testing.
- Refactored various components to maintain float precision in settings, curves, and tables, ensuring accurate data representation.
- Implemented label expression evaluations for dynamic indicator labels in rusEFI.
- Added comprehensive tests for float values and label expressions to ensure robustness and correctness.

+++ packages/foxtune_tune/test/real_tables_test.dart
@@ -110,7 +110,7 @@ void main() {
     final tune = TuneState.empty(doc);
     final algorithm = tune.locate('algorithm')!;
     // Select TPS as the load source; the axis label must follow.
-    final current = tune.readRaw(algorithm.page, algorithm.field)!;
+    final current = tune.readRaw(algorithm.page, algorithm.field)!.toInt();
     final field = algorithm.field as IniBitsField;
     final mask =
         ((1 << (field.highBit - field.lowBit + 1)) - 1) << field.lowBit;
